added stripe connect integration for customers to take payments from their clients

+++ src/main/java/com/nexus/expense/CreateExpenseRequest.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.Positive;
 import java.time.Instant;
 
 public record CreateExpenseRequest(
-        @Positive double amount,
+        @Positive long amount,
         Instant paymentDate,
         Integer projectId,
         @Positive Long expenseCategoryId
