+++ src/main/java/com/nexus/expense/Expense.java
@@ -14,13 +14,13 @@ public class Expense extends AbstractPayment {
     @ManyToOne(optional = false)
     private ExpenseCategory expenseCategory;
 
-    public Expense(double amount, String currency, Instant paymentDate, Project project, ExpenseCategory expenseCategory, UUID tenantId) {
+    public Expense(long amount, String currency, Instant paymentDate, Project project, ExpenseCategory expenseCategory, UUID tenantId) {
         super(amount, paymentDate, project, currency);
         this.expenseCategory = expenseCategory;
         setTenantId(tenantId);
     }
 
-    public Expense(double amount, String currency, Instant paymentDate, ExpenseCategory expenseCategory, UUID tenantId) {
+    public Expense(long amount, String currency, Instant paymentDate, ExpenseCategory expenseCategory, UUID tenantId) {
         super(amount, paymentDate, currency);
         this.expenseCategory = expenseCategory;
         setTenantId(tenantId);
