+++ src/main/java/com/nexus/expense/UpdateExpenseRequest.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.Positive;
 import java.time.Instant;
 
 public record UpdateExpenseRequest(
-        @Positive double amount,
+        @Positive long amount,
         Instant paymentDate,
         @Positive Long expenseCategoryId
 ) {
