+++ src/main/java/com/nexus/income/BasicIncomeResponse.java
@@ -2,6 +2,6 @@ package com.nexus.income;
 
 public record BasicIncomeResponse(
         Long id,
-        double amount,
+        long amount,
         String paymentDate
 ) { }
