+++ src/main/java/com/nexus/income/UpdateIncomeRequest.java
@@ -5,6 +5,6 @@ import jakarta.validation.constraints.Positive;
 import java.time.Instant;
 
 public record UpdateIncomeRequest(
-        @Positive double amount,
+        @Positive long amount,
         Instant paymentDate
 ) { }
